feat: Pre-consumption script to decode encrypted pdfs
Relies on passwords.txt which the user has to fill/supply themselves,
though it's prefilled with some of the most common passwords that have
been found in various data dumps.
No cracking is performed. The script will try them out one by one until
one matches or it runs out of passwords to try.
If the password is over a certain length, the script will output a subtle
reminder with the character length and first and last character of the
matched password.

CAVEAT: the original document version will have no encryption
afterwards, if that is important to you.

Refer-to: https://en.wikipedia.org/wiki/List_of_the_most_common_passwords and other language variants
Refer-to: https://github.com/tarraschk/richelieu
Refer-to: https://nordpass.com/de/most-common-passwords-list/